pyproject.toml
[tool.poetry]
name = "qiskit-qutrit-calibration"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
qiskit-experiments = "^0.5.4"
qiskit-ibm-provider = "^0.8.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
stestr = "^4.1.0"
astroid = "^3.0.2"
pylint = "^3.0.3"
sphinx = "^7.2.6"
qiskit-sphinx-theme = "^1.16.0"
jupyter-sphinx = "^0.5.3"
pygments = "^2.17.2"
reno = "^4.0.0"
tox = "^4.11.4"
setuptools = "^69.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
